+++ src/main/java/com/app/service/booking/BookingService.java
@@ -12,7 +12,7 @@ public interface BookingService {
     Booking saveOrUpdate(Booking booking);
     Booking getById(Long id);
     BookingNote addNote(Long userId, Long bookingId, String note);
-    List<BookingNote> getNotes(Long bookingId, Integer offset);
+    List<BookingNote> getNotes(Long bookingId, Long offset);
     List<Booking> getForDates(Date dateFromMin, Date dateFromMax, Date dateToMin, Date dateToMax);
     List<Booking> getExistingBookings(Long userId, Date dateFrom, Date dateTo);
     Booking getByReference(String reference);
